+++ Source/Modules/ISHServer/Private/Get-ISHFTPItem.ps1
@@ -14,21 +14,33 @@
 # limitations under the License.
 #>
 
-function Set-ISHFirewallOracle
+function Get-ISHFTPItem
 {
     [CmdletBinding()]
-    Param()
+    param(
+        [Parameter(Mandatory=$true)]
+        [string]$FTPHost,
+        [Parameter(Mandatory=$true)]
+        [pscredential]$Credential,
+        [Parameter(Mandatory=$true)]
+        [string[]]$Path,
+        [Parameter(Mandatory=$true)]
+        [string]$LocalPath
+    )
     
     begin 
     {
-        . $PSScriptRoot\Private\Test-RunningAsElevated.ps1
-        Test-RunningAsElevated -StopCallerPSCmdlet $PSCmdlet
+        Import-Module PSFTP -ErrorAction Stop
+        Set-FTPConnection -Server $FTPHost -Credentials $Credential -UseBinary -KeepAlive -UsePassive | Out-Null
     }
 
     process
     {
-        $ruleNameTCP="ISH - Oracle - TCP"
-        # http://docs.sdl.com/LiveContent/content/en-US/SDL%20Knowledge%20Center%20full%20documentation-v2/GUID-4A04043C-FAC7-4839-AB7E-1E29C9CA64F7
+        $Path | ForEach-Object {
+            Write-Debug "Path=$_"
+            Get-FTPItem -Path $_ -LocalPath $LocalPath -Overwrite | Out-Null
+            Write-Verbose "Downloaded $_ to $LocalPath"
+        }
     }
 
     end