+++ Source/Modules/ISHServer/Set-ISHFirewallMSDTC.ps1
@@ -21,7 +21,7 @@ function Set-ISHFirewallMSDTC
     
     begin 
     {
-        . $PSScriptRoot\Test-RunningAsElevated.ps1
+        . $PSScriptRoot\Private\Test-RunningAsElevated.ps1
         Test-RunningAsElevated -StopCallerPSCmdlet $PSCmdlet
     }